# Time optimization in do_photometry